cotation/requirements.txt
Flask==3.1.1
schedule==1.2.0
python-binance==1.0
plotly==6.1.1
numpy==2.2
polars==1.30
pyarrow==20.0
jsonify==0.5
psutil==7.0
matplotlib==3.10
mplfinance==0.12.10b0
